+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.85.3
 
+- backups: Add bioText & bioEmoji fields to AccountData